Include the EIP191 message prefix as part of message hash
+ The parameters that are being hashed for the message now include a
prefix hash = '\x19Ethereum Signed Message:\n' + message.length

This still conforms to the EIP191 standard
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-191.md

However this method lets us only perform a single keccak256 hash that
includes the prefix and all parameters instead of hashing the parameters
and then wrapping it in another hash to include the prefix. Also a side
effect of this change is the sighash that is emitted in the event of the
escrow is now exactly the same sighash that is verified with the
ECDSA.verify call.

+ The message hash now also includes a message type id to differentiate
between different messages that could be sent to a single escrow
contract