Fix unit tests to use the correct sighash when signing messages
+ This requires the js SigningService to abi.econdePacked() all the
parameters of the message along with the message prefix and then hash
and sign it. Web3.Eth.Sign prepends the prefix to any message you give
it and then hashes and signs the message without doing the packing for
you. The method soliditySha3() will do the correct packing for you
before signing a message but will also hash the message. In order to get
it to just encode the parameters and not hash it requires using methods
that are not publically exposed through the public web3 package so
unforunately I had to directly include the code in this project.

The code was taken from

+ https://github.com/ethereum/web3.js/blob/2.x/packages/web3-utils/src/Utils.js
+ https://github.com/ethereum/web3.js/blob/2.x/packages/web3-utils/src/SoliditySha3.js

There is also a issue open on the web3.js project requesting as a
feature publically exposing these methods for this very purpose:

https://github.com/ethereum/web3.js/issues/2541