Escrow Library Contract
This abstracts the escrow payment channel contract into 2 separate
layers to decouple as much of the common business logic needed for
escrow contracts into a common library contract but still having each
escrow be a unique contract address on the blockchain that holds the
escrowed funds.

The main motivation behind this refactor is to make each escrow that is
created (by the factory contract) as thin as possible in order to reduce
gas costs of deploying new contracts which is currently the majority of
the gas cost of opening/closing an escrow payment channel.

A separate EscrowLibrary contract is now responsible for containing
all the business logic and escrow mutable/unmutable state. This works
by storing a mapping of escrow address to escrow parameters (also
puzzle parameters if a puzzle is posted for that escrow).

All escrow events our now emitted in the context of the EscrowLibrary
contract so that ethereum clients can subscribe to logs on just a
single contract while still filtering on the particular escrow address
as an indexed topic in the event.

The actual transfer of funds is still exectuted in the context of the
escrow contract because this is where the funds are sent. Therefore
this logic must be delegated to the escrow contract via a send method
on the escrow contract that is guarded so only the EscrowLibrary
contract can initiate a transfer.

The ETH Escrow contract still contain a payable fallback method that calls the
library contract to emit an EscrowFunded event when it is funded.

EthEscrow and Erc20Escrow contracts only implement 2 abstract methods:

+ balance - returns the current balance (ether or token balance)
+ send - sends ether or token to the provided address (only callable
from EscrowLibrary)