Escrow Library Contract
This abstracts the escrow contract into a 2 contract layer system to
decouple as much of the common business logic of escrow contracts from
the pure storage/transfer logic that must be unique for each escrow.

The main motivation behind this refactor is to make each escrow that is
created (by the factory contract) as thin as possible in order to reduce
gas costs of deploying new contracts which is currently the majority of
the gas cost of opening/closing a escrow payment channel.

A separate escrow library contract is now responsible for containing all
the business logic of checking prerequisite state conditions, computing
sighashes from inputs and verifying signatures. It does this by taking a
given escrow contract as an additional parameter to each method and
looking up all relevant state information from public getters on the
escrow. In addition now all escrow events our now emitted in the context
of the EscrowLibrary contract so that ethereum clients can subscribe to
logs on just a single contract while still filtering on the particular
escrow address as an indexed topic in the event.

In addition it stores the PuzzleParams for a given escrow in a
mapping. This is done so a wrapper for setting puzzle parameters within
the escrow itself is not needed. Note this can be easily extended to
support bidirectionality in the future to contain a sequence number and
allow for a challenge period where greater sequence number puzzles to
posted game.

Whenever ether/ERC20 token transfer it is still exectuted in the context
of the escrow (only a contract can modify its own state or transfer
ether it holds). Therefore this logic must be delegated to the escrow
contract via a public method on the escrow contract. In addition the
escrow currently needs to know its escrow state so it can check its
state in the pyable fallback method of the Eth Escrow.

The set of methods currently available only to the library contract include:

+ function setState(EscrowState escrowState) public;
+ function closeEscrow() public;
+ function sendToEscrower(uint _amt) public;
+ function sendRemainingToEscrower() public;
+ function sendToPayee(uint _amt) public;
+ function sendRemainingToPayee() public;

The ETH Escrow contract still contain a payable fallback method that calls the
library contract to emit an EscrowFunded event when it is funded.

The ERC20 Escrow contract contains a method that attempts to transfer an
ERC20 token (which must have been previously approved) into the
escrow. The library contains a wrapper of this method which also emits
an EscrowOpened event.