[EscrowLibrary] Use withdraw pattern and check return value of send
+ Adds a `withdraw` function to be used in cases where an escrow is
closed with a puzzle and solve or puzzle and puzzleRefund to prevent a
malicious reserveAddress that always reverts from trapping funds in
the escrow. In these cases after closing the escrow with a
solve/puzzleRefund the escrower and payee must call withdraw to pull
funds out of the escrow.

+ Adds sendEscrower and sendPayee functions to the
EscrowLibrary. Check the return value of each Escrow.send call and
only zero out the escrower/payee balance in the case of a successful
send. Emit a `FundsTransferred` event with this return value so that
sends do not fail silently