[EscrowLibrary] Fix sendEscrower/sendPayee reentrancy
Zero-out the escrower/payee balance before sending in sendEscrower and
sendPayee functions. Revert if the send fails. This should be safe to
do now that the payee can always post the puzzle and withdraw
separately from the escrower when closing an escrow.

Now that send always reverts in case of a failure it is no longer
necessary to include the send success boolean in the FundsTransferred
event.