[Escrow, Library, Factory] Use CREATE2 for deterministic addresses
+ Use CREATE2 to create escrows at a deterministic address using the
escrowParamsHash as salt and hard coded escrow contract bytecode. This
eliminates issues with relying on the funding address of the escrow
not changing once the escrow params have been agreed on.

+ Simplify logic for opening escrows since we can rely on the escrow
address being prefunded before the call to the factory that creates
the escrow. Therefore the create function on the factory can just
trigger the opening logic in the EscrowLibrary when a new escrow is
being created. This removes the need for a checkFunded function or a
EscrowFunded/EscrowOpened events. Finally the EthEscrow payable
fallback can be removed because now its assumed all funds will be sent
to the address prior to its creation.

+ Remove the escrow params mapping in the factory as now the create2
call will fail if a duplicate escrow is created.