[Solc] Adds an external truffle compile script
compile-constracts.sh is used to standardize the compiler
version/flags and metadata that is generated when compiling some of
the smart contracts. This is needed for EthEscrow.sol and
ERC20Escrow.sol because we hard-code the bytecode of these conracts in
the EscrowFactory in order to use the create2 opcode.